Fixed bug in mdtable_csv_file where a i_ was being appended to the first column name
Problem: On Windows machines if you do not specify the file encoding to
be UTF-8-BOM when reading in a CSV file a i.. is appended to the first
column header name. For example, if the column name is measures it will
become i..measures.

Solution: Add the fileEncoding argument when using the read.csv
function.